+++ examples/complete/outputs.tf
@@ -8,12 +8,12 @@ output "vpc_details" {
   description = "Details of all created VPCs"
   value = {
     for k, v in module.iter_terraform_complete : k => {
-      vpc_id             = try(v.vpc_id, null)
-      vpc_cidr_block     = try(v.vpc_cidr_block, null)
-      private_subnets    = try(v.private_subnets, [])
-      public_subnets     = try(v.public_subnets, [])
-      database_subnets   = try(v.database_subnets, [])
-      nat_gateway_ids    = try(v.natgw_ids, [])
+      vpc_id           = try(v.vpc_id, null)
+      vpc_cidr_block   = try(v.vpc_cidr_block, null)
+      private_subnets  = try(v.private_subnets, [])
+      public_subnets   = try(v.public_subnets, [])
+      database_subnets = try(v.database_subnets, [])
+      nat_gateway_ids  = try(v.natgw_ids, [])
     } if can(v.vpc_id)
   }
 }
@@ -38,4 +38,26 @@ output "secrets_details" {
     } if can(v.secret_id)
   }
   sensitive = true
+}
+
+output "kms_key_details" {
+  description = "Details of all created KMS keys"
+  value = {
+    for k, v in module.iter_terraform_complete : k => {
+      key_id  = try(v.key_id, null)
+      key_arn = try(v.key_arn, null)
+      aliases = try(v.aliases, {})
+    } if can(v.key_id)
+  }
+}
+
+output "dynamodb_table_details" {
+  description = "Details of all created DynamoDB tables"
+  value = {
+    for k, v in module.iter_terraform_complete.dynamodb_tables : k => {
+      table_id   = try(v.dynamodb_table_id, null)
+      table_arn  = try(v.dynamodb_table_arn, null)
+      stream_arn = try(v.dynamodb_table_stream_arn, null)
+    }
+  }
 }